api/requirements.txt
# Core FastAPI and server
fastapi[standard]
uvicorn[standard]

# AI/ML Model APIs
openai
together
anthropic
google-genai
xai-sdk
instructor

# Data validation and settings
pydantic
pydantic-settings

# Database and authentication
supabase
pyjwt
requests
httpx

# Data processing and analysis
numpy
pandas
nltk
scipy
scikit-learn
tiktoken
deepeval

# Utilities
typing-extensions